% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FpkmTpmCount.R
\name{countToTpm}
\alias{countToTpm}
\title{TPM FPKM COUNT conversion}
\usage{
countToTpm(counts, effLen = 5)
}
\arguments{
\item{counts}{raw count of RNAseq data.}

\item{fpkm}{FPKM.}
}
\value{
generate a expression matrix (TPM or FPKM).
}
\description{
TPM FPKM COUNT conversion
}
\examples{
tpm <- countToTpm(counts = counts)
fpkm <- countToFpkm(counts = counts)
tpm <- fpkmToTpm(counts = fpkm)
}
\keyword{dataConversion}
