% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLimmaOutput.R
\name{createLimmaOutput}
\alias{createLimmaOutput}
\title{using limma package to do differential expression analysis}
\usage{
createLimmaOutput(
  GSE = "GSE65858",
  columnGSE = 1:ncol(eset),
  GPL = "GPL10558",
  method = "max",
  columnGPL = NULL,
  group = NULL,
  sepTitle = "_",
  groupIndex = 5,
  logFC = 1.5,
  pvalue = 0.05
)
}
\arguments{
\item{GSE}{a GSE id or a data.frame. if a data.frame, first column must be gene symbol.}

\item{columnGSE}{which column should be used to do differential expression analysis.}

\item{GPL}{the matched GPL platform of GSE id.}

\item{method}{max or mean.}

\item{columnGPL}{which column to convert probe id.}

\item{group}{if GSE is a data.frame, a group vector should be provided. Otherwise omit.}

\item{sepTitle}{separator of the title.}

\item{groupIndex}{which vector should be taken after the title was split.}

\item{logFC}{log fold change.}

\item{pvalue}{cutoff to screen result.}
}
\value{
generate a differential expression analysis results.
}
\description{
using limma package to do differential expression analysis
}
\examples{
createLimmaOutput(GSE = fpkm, group = group)
createLimmaOutput(GSE="GSE19136",columnGSE = c(1,2,4,5,7,8,10,11),GPL='GPL570', method="max", columnGPL=c(1,11),logFC=1.5,pvalue=0.05)
}
\keyword{limmaResult}
