% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classPredict.R
\name{plotROCCurve}
\alias{plotROCCurve}
\title{Plot ROC curves}
\usage{
plotROCCurve(list, method)
}
\arguments{
\item{list}{list returned by function \code{classPredic}.}

\item{method}{character string of a prediction method.
\itemize{
   \item "ccp":  compound covariate predictor
   \item "dlda": diagonal linear discriminant analysis
   \item "bcc":  Bayesian compound covariate predictor
}}
}
\description{
This function plots ROC curves for compound covariate predictor,
diagonal linear discriminant analysis and Bayesian compound covariate predictor methods.
}
\examples{
res <- test.classPredict("Brca")
plotROCCurve(res, "ccp")
plotROCCurve(res, "dlda")
plotROCCurve(res, "bcc")
}
